fix(api-gw): Fixed integration tests to avoid DB mock connection, and removed hash domain to use link domain.
