+++ packages/api-gw/package.json
@@ -12,7 +12,7 @@
   "scripts": {
     "start": "node src/index.js",
     "start:dev": "nodemon src/index.js",
-    "test": "NODE_ENV=test jest --config ./config/jest.config.js --watch-all",
+    "test": "jest --config ./config/jest.config.js --watch-all",
     "coverage": "jest --coverage",
     "inspect": "node --inspect src/index.js"
   },
@@ -27,7 +27,6 @@
     "fastify-plugin": "^3.0.0"
   },
   "devDependencies": {
-    "@shelf/jest-mongodb": "^1.2.3",
     "jest": "^26.6.3",
     "nodemon": "^2.0.7"
   }
