+++ packages/api-gw/src/adapters/healthyChecks/healthyChecks.integration.js
@@ -1,18 +1,28 @@
 const fastify = require('fastify');
-const { buildMockServer } = require('../../utils/testUtils');
-const statusRouter = require('../../delivery/http/routes/status');
+const { createReqStub, createReplyStub } = require('../../utils/testUtils');
 
 describe("/status", () => {
   let server;
 
   beforeEach(async () => {
-    server =  fastify({});
-    server.register(statusRouter);
+    server = fastify({});
+    const livenessMock = jest.fn(() => ({ status: 'Server Running' }))
+    server.get('/status', livenessMock);
     await server.ready();
   });
 
+  afterEach(async function () {
+    if (server) {
+      await server.close();
+      server = null;
+    }
+  })
+
   it("GET: Given a request then it must return running", async () => {
-    const response = await server.inject({ method: "GET", url: "/" });
+    // Act
+    const response = await server.inject({ method: "GET", url: "/status" });
+
+    // Assert
     expect(response.statusCode).toEqual(200);
     expect(response.json().status).toEqual('Server Running');
   });
