+++ packages/api-gw/src/adapters/index.js
@@ -1,9 +1,7 @@
 const healthyChecks = require('./healthyChecks/healthyChecks');
 const link = require('./link/link');
-const hash = require('./hash/hash')
 
 module.exports = {
   healthyChecks,
-  link,
-  hash
+  link
 };
