+++ packages/api-gw/src/adapters/link/link.spec.js
@@ -28,4 +28,56 @@ describe('Link Adapters' , () => {
       expect(((replyStubs.code()).headers()).send).toHaveBeenCalled();
     });
   });
+
+  describe('redirectToUrl', () => {
+    it('Given a valid hash the function must redirect to the resulting url', async () => {
+      // Arrange
+      const reqStubs = createReqStub({
+        params: {
+          hash: Faker.random.uuid()
+        }
+      });
+      const replyStubs = createReplyStub();
+      const linkAdapterStub = {
+        readUrlByHash: jest.fn(() => Promise.resolve({ originalUrl: Faker.internet.url() })),
+      }
+
+      // Act
+      const readUrlByHashBinded = linkAdapters.redirectToUrl
+        .bind({ linkAdapter: linkAdapterStub });
+      await readUrlByHashBinded(reqStubs, replyStubs);
+
+       // Asserts
+       expect(reqStubs.log.info).toHaveBeenCalled();
+       expect((replyStubs.redirect)).toHaveBeenCalled();
+    })
+
+    it('Given an invalid hash the function must return a 404 error', async () => {
+       // Arrange
+       const reqStubs = createReqStub({
+        params: {
+          hash: Faker.random.uuid()
+        }
+      });
+      const replyStubs = createReplyStub();
+      const linkAdapterStub = {
+        readUrlByHash: jest.fn(() => Promise.resolve(null)),
+      }
+
+      // Act
+      const readUrlByHashBinded = linkAdapters.redirectToUrl
+        .bind({ linkAdapter: linkAdapterStub });
+      await readUrlByHashBinded(reqStubs, replyStubs);
+
+      // Asserts
+      expect(reqStubs.log.info).toHaveBeenCalled();
+      expect(replyStubs.code).toHaveBeenCalledWith(404);
+      expect((replyStubs.code()).headers).toHaveBeenCalled();
+      expect(((replyStubs.code()).headers()).send).toHaveBeenCalledWith({
+        statusCode: 404,
+        error: "Not found",
+        message: "Oops, it looks like this url do not exist anymore"
+      });
+    })
+  });
 });
