+++ packages/mongodb-connection-module/src/index.js
@@ -6,21 +6,13 @@ const logger = require('./utils/logger');
 
 Mongoose.Promise = global.Promise;
 
-function createDb() {
-  if (process.env.NODE_ENV === supportedEnvs.TEST) {
-    return Mongoose.createConnection(global.__MONGO_URI__)
-  }
-
-  return Mongoose.createConnection(config.mongodb.uri, {
-    useNewUrlParser: true,
-    auth: {
-      user: config.mongodb.user,
-      password: config.mongodb.pass,
-    },
-  });
-}
-
-const db = createDb()
+const db = Mongoose.createConnection(config.mongodb.uri, {
+  useNewUrlParser: true,
+  auth: {
+    user: config.mongodb.user,
+    password: config.mongodb.pass,
+  },
+});
 
 db.on('error', (err) => {
   logger.error(`[pl-mongodbconnection-module]: Connection error event: ${err.message}`);
